ftrace: make FENTRY_RECORD independent
We are going to use mcount record in kgr along with dynamic ftrace. So
extract the common code under 2 config options:
* FENTRY_RECORD -- mostly what FTRACE_MCOUNT_RECORD used to be. This
  means we want to actually generate mcount/fentry calls. Used by both
  static and dynamic ftrace.
* FENTRY_RECORD_LIB -- we dynamically patch away theses calls to be
  nops, i.e. this is what dynamic ftrace does and what kgr will use
  later too. Static ftrace is not compatible with this as it patches
  dynamicall nothing and has to have the calls in place.

So this results in three possible configurations:
* dynamic_ftrace and kgr enabled (selects FENTRY_RECORD +
  FENTRY_RECORD_LIB)
* ftrace disabled completely and kgr enabled (selects FENTRY_RECORD +
  FENTRY_RECORD_LIB)
* static_ftrace enabled; dynamic_ftrace with kgr disabled
  (selects only FENTRY_RECORD)

One exception is FENTRY_RECORD_LIB around the ftrace_init declaration
in include/linux/ftrace.h which is switched to DYNAMIC_FTRACE, the
same as the ftrace_init definition in kernel/trace/ftrace.c.

Signed-off-by: Jiri Slaby <jslaby@suse.cz>