ftrace: Clean ftrace_disabled handling in ftrace_update_code
Check the returns directly and return from ftrace_update_code right
when an error happens instead of waiting for the next loop.

This makes the flow of the code natural and no games are played behind
the reader's mind. ftrace_disabled is set to 1 at the root of call
chains.

ftrace_disabled is not needed in ftrace_code_disable now, as we exit
from the loop at the right time. We can do all this because
ftrace_disabled is checked in ftrace_init_module too and ftrace_init
is the first to call in the ftrace code, so that ftrace_disabled
cannot be set yet.

Signed-off-by: Jiri Slaby <jslaby@suse.cz>