kgr: allow replace_all
We may want load a single patch on the top of others which will
replace all of them atomically. As a result we have only a single
patch applied and all previous modules can be unloaded.

It works as follows:
* all patches and its patch_funs are walked
* if a patch_fun's function is contained in the new patch, it is
  moved to the new one, using the slow stub
* if it is not, the patch_fun is reverted

This way, all patch_funs are reverted sequentially. After parallel
patching is implemented, we can get rid of the kgr_to_revert list and
dump all the previous patches to be reverted just to the running
patching.

pm: clean up reference handling when a patch replaces all others
pm: correctly remove replaced patch when using fast path
mb: fix replace_all {
Current implementation of replace_all functionality does not fully allow
to go back in the patch series. The situation can be illustrated in the
following way.

Kernel contains functions A and B. Patch1 patches both functions (new
versions are called A1 and B1). Patch2 (replace_all) only function A
(A2). If A calls B there is a race window during application of Patch2.
A2 can call B1, which is wrong. It should call the original B. It is
caused by postponed revert of B1 after finalization of A2.

Also two-step finalization of the replace_all patch is wrong as it leads
to bugs. First the "forward" patch_funs of the patch (A) are finalized.
Then reverted functions (B) have to be finalized after another
kgr_modify_kernel call. This second step however includes also another
finalization of the forward functions which is not possible. Finally,
refs in kgr_patch structures are not correct.

This fix preserves the original idea of the replace_all process, but
implements it earlier in kgr_modify_kernel.

* First, we walk through previous patches in kgr_patches list and their
  functions patch_fun. If patch_fun is not present in replace_all patch
  it is reverted (slow revert). Nothing is done in the opposite case.
* Then all patches for functions in the replace_all patch are applied.
  There is no race window, because we are protected by immutable flag.

Finalization is very similar.

* First, all patch_funs from the replace_all are finalized in the
  ordinary way.
* Then we can finalize the reverted patch_funs from previous patches.
  Non reverted patch_funs are pushed right to the KGR_PATCH_REVERTED
  state ("fast revert").
* Next all patches from kgr_patches list can be removed, because they
  are no longer used (and fast stubs are already applied).

Apart from these changes consistency with stacking of patches has to be
preserved.

* In case of reverted patch_funs (i.e. not present in replace_all patch,
  e.g. B1) ftrace needs to be update only for last patch_funs for each
  function (i.e.  for B1 and not some hypothetical B0 applied
  beforehand).  This applies to APPLIED and REVERT_SLOW states in
  kgr_patch_code.
* In slow stub, calls to reverted patch_funs in new universe have to be
  redirected to the original functions. I.e. if we have series of
  stacked patch_funs B0->B1 over the original B, calls to B must really
  call B and not B0 in the new universe (however it has to call B1 in
  the old one). Instead of editing the slow stub we change loc_old to
  loc_name before the actual revert is started.
* After finalization loc_old of forward patch_funs have to be switched
  to loc_name to prepare correct revert of the replace_all patch (if
  required). It cannot be done during initialization of ftrace_ops
  because we need correct loc_old (in terms of stacking) for the slow
  stub and old universes.

Finally, delayed loading of modules has to be updated. Some module with
to be patched functions can be loaded while the replace_all patch is
being applied. We need to get to the same state as if the module had
been loaded before the replace_all patch came.

}
mb: handle the failure of finalization stage
mb: add error code to the message in kgr_revert_replaced_funs
pm: set revert slow state for all reverted symbols when loading patched module

Signed-off-by: Miroslav Benes <mbenes@suse.cz> [fixes]
Signed-off-by: Petr Mladek <pmladek@suse.cz> [fixes]
Signed-off-by: Jiri Slaby <jslaby@suse.cz>