kgr: remove abort_if_missing flag
abort_if_missing flag is intended to inform kgraft if it should abort
patching when the function is not found in current environment.
Typically set to 'true' for vmlinux functions and to 'false' for module
functions. We now have objname specified for each function and thus we
can remove the flag from the code. Object name is more suitable for the
task anyway.

This also allows us to remove FIXME in
kgr_handle_patch_for_going_module(). It is closely connected with
abort_if_missing semantics. Objname has a different semantics (it is not
the exact opposite of the flag) so we do not need to do anything when
the module goes away.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>