kgr: fix subtle race with kgr_module_init(), going notifier and kgr_modify_kernel()
This is a modified backport of upstream commit 8cb2c2dc4727 ("livepatch:
Fix subtle race with coming and going modules"). Quoting Petr:

"There is a notifier that handles live patches for coming and going
modules. It takes klp_mutex lock to avoid races with coming and going
patches but it does not keep the lock all the time. Therefore the
following races are possible:

      1. The notifier is called sometime in STATE_MODULE_COMING. The
module is visible by find_module() in this state all the time. It means
that new patch can be registered and enabled even before the notifier is
called. It might create wrong order of stacked patches, see below for an
example.

       2. New patch could still see the module in the GOING state even
after the notifier has been called. It will try to initialize the
related object structures but the module could disappear at any time.
There will stay mess in the structures. It might even cause an invalid
memory access.

    This patch solves the problem by adding a boolean variable into
struct module.  The value is true after the coming and before the going
handler is called.  New patches need to be applied when the value is
true and they need to ignore the module when the value is false."

The same applies also to kGraft although we do not have a coming
notifier. It is replaced with kgr_module_init() function called directly
from load_module(). Thanks to a different implementation the race as
described in point 1 above is in fact harmless. The race could lead to
a double patching of a function but there is a check for
KGR_PATCH_SKIPPED state in kgr_module_init() which prevents. However
we're gonna remove the check so the consequences might be more serious.
The second race is valid nevertheless.

Originally-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>