kgr: add sympos as disambiguator field to kgr_patch_fun structure
Currently, it is not possible to patch functions with ambiguous names.
kallsyms_lookup_name() always returns only the first function address it
finds. There is of course a workaround to patch all its calling
functions but this can get ugly easily. We can use symbol's occurrence
in kallsyms table to disambiguate between them.

To quote Chris:
"In cases of duplicate symbols, sympos will be used to disambiguate. By
default sympos will be 0, and patching will only succeed if the symbol
is unique. Specifying a positive value will ensure that occurrence of
the symbol in kallsyms for the patched object will be used for patching
if it is valid."

Moreover, introduce a new macro which allow to specify sympos.

Originally-by: Seth Jennings <sjenning@redhat.com>
Originally-by: Chris J Arges <chris.j.arges@canonical.com>
Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>