kgraft/bcache: Do not block livepatching in the writeback thread for too long
The writeback kthread waits 30 sec by default when the full search
has finished. The sleeping cycle is a safe location for the migration
and we should do so there.

There is a similar sleeping cycle in the read_dirty() function.
But it is not clear that it is a safe location for the migration.
There is called bch_keybuf_next() before. Fortunately, the delay
is one second at maximum, so we could leave it as is.

On the other hand, the main cycle calls try_to_freeze() only
when there are no dirty data. It might take some time until
it happens. And it should be safe to migrate the kthread in
every cycle before calling down_write(). At least, it should
not be more dangerous than migrating the thread during the
30 sec long sleep or inside the read_dirty() call.

Reported-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>